+++ NEWS.md
@@ -1,3 +1,7 @@
 # triact 0.2.0
 
 * First version on CRAN
+
+# triact 0.3.0_dev
+
+* fix argument check for timeFwdUpRight_cols of Triact$load_files()